set up es6 project structure with build systems
