+++ .gitignore
@@ -1,3 +1,3 @@
 meyda.sublime-project
 meyda.sublime-workspace
-LICENSE
+node_modules
