+++ Makefile
@@ -1,0 +1,5 @@
+build:
+	@./node_modules/.bin/gulp
+
+test: build
+	@./node_modules/.bin/mocha test
