+++ bower.json
@@ -1,6 +1,6 @@
 {
   "name": "meyda",
-  "main": "meyda.min.js",
+  "main": "lib/meyda.min.js",
   "version": "1.1.0",
   "homepage": "https://github.com/hughrawlinson/meyda",
   "authors": [
