+++ src/extractors/energy.js
@@ -1,0 +1,7 @@
+export default function(bufferSize, m) {
+  var energy = 0;
+  for(var i = 0 ; i < m.signal.length ; i++){
+    energy += Math.pow(Math.abs(m.signal[i]),2);
+  }
+  return energy;
+}
