+++ src/extractors/extractorUtilities.js
@@ -1,0 +1,9 @@
+export function mu(i, amplitudeSpect){
+  var numerator = 0;
+  var denominator = 0;
+  for(var k = 0; k < amplitudeSpect.length; k++){
+    numerator += Math.pow(k,i)*Math.abs(amplitudeSpect[k]);
+    denominator += amplitudeSpect[k];
+  }
+  return numerator/denominator;
+}
