+++ src/extractors/spectralCentroid.js
@@ -1,0 +1,5 @@
+import mu from 'extractorUtilities';
+
+export default function(bufferSize, m){
+  return mu(1,m.ampSpectrum);
+}
